Adding tests for data component
Added tests for data components. To achieve this, we had to alter the structure of the files. Any changes would need to accomdate this new file structure. Primary code is contained in src directory and testing scripts are contained in tests directory. The __init__.py in src exposes functionality to the tests directory.